tons of refactoring
