+++ src/components/ModalItemListItem.jsx
@@ -1,0 +1,41 @@
+import React from "react";
+import PropTypes from "prop-types";
+
+import ItemIcon from "./ItemIcon";
+import ItemData from "./ItemData";
+import ItemUtility from "../utility/ItemUtility";
+import PropTypeUtility from "../utility/PropTypeUtility";
+
+export default class ModalItemListItem extends React.Component {
+    getLevel() {
+        if(this.props.type === "Lantern") {
+            return 0;
+        }
+
+        const itemType = ItemUtility.itemType(this.props.item.type);
+
+        return ItemUtility.maxLevel(itemType === "Weapon" ? "weapons" : "armours",
+            this.props.item.name);
+    }
+
+    render() {
+        const item = this.props.item;
+        const type = this.props.type;
+        const level = this.getLevel();
+
+        return <div className="item-title-wrapper">
+            <div className="item-wrapper">
+                <div className="item" title={item.description} onClick={() => this.props.onSelected(type, item.name)}>
+                    <ItemIcon item={item} defaultType={ItemUtility.itemType(this.props.item.type)} />
+                    <ItemData item={item} level={level} renderCellLine={true} />
+                </div>
+            </div>
+        </div>;
+    }
+}
+
+ModalItemListItem.propTypes = {
+    type: PropTypes.string,
+    item: PropTypeUtility.item(),
+    onSelected: PropTypes.func
+};
