+++ src/utility/ItemUtility.jsx
@@ -19,4 +19,34 @@ export default class ItemUtility {
 
         return Math.max(...Object.keys(DataUtility.data()[collection][itemName][levelKey]).map(k => Number(k)));
     }
+
+    static levelString(level = 0) {
+        let levelString = "";
+
+        if(level && level > 0) {
+            levelString = `+${level}`;
+        }
+
+        return levelString;
+    }
+
+    static itemType(type) {
+        switch(type) {
+            case "Weapon":
+            case "Sword":
+            case "Chain Blades":
+            case "Axe":
+            case "Hammer":
+            case "War Pike":
+            case "Repeater":
+                return "Weapon";
+            case "Head":
+            case "Torso":
+            case "Arms":
+            case "Legs":
+                return "Armour";
+        }
+
+        return "Lantern";
+    }
 }
